Move GroupedProviderCoursesComponent under candidate namespace
