+++ app/components/candidate_interface/grouped_provider_courses_component.html.erb
